+++ 04.class/src/MemoModel.js
@@ -1,0 +1,9 @@
+export class MemoModel {
+  constructor(content) {
+    this.body = content.body;
+  }
+
+  getTitle() {
+    return this.body.split("\n")[0];
+  }
+}
